pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "halo"
version = "0.1.0"
description = "Heterarchical Associative Learning Orchestration — a biologically inspired cortical learning framework"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.9",
]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
